mysql_loader/requirements.txt
pymongo
requests
pandas
pyarrow
mysql-connector-python